quizs/requirements.txt
aiokafka==0.7.2
amqp~=5.0.9
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
click~=8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-celery-beat==2.2.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-redis==5.0.0
django-timezone-field==4.2.3
djangorestframework==3.12.4
kafka-python==2.0.2
kombu~=5.2.3
mmh3==3.0.0
pottery==2.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.2
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
super-logger==0.1.4
supervisor==4.2.2
typing_extensions==4.1.1
vine==5.0.0
wcwidth==0.2.5
openpyxl==3.0.9
--extra-index-url http://s.dev.isgneuro.com/repository/ot.platform/simple
--trusted-host s.dev.isgneuro.com